worked on speed up and slow down - speed up works - slow down has a bug

+++ Bricks/src/Game.java
@@ -286,7 +286,7 @@ public class Game extends JPanel implements MouseListener {
 		while (true) {
 			game.move();
 			game.repaint();
-			if (placeHolder.isEmpty() == false){
+			if (placeHolder.isEmpty() == false){ // placeholder is powerups on screen
 				for (int j=0; j<placeHolder.size();j++){
 					placeHolder.get(j).move();
 				}
@@ -331,11 +331,11 @@ public class Game extends JPanel implements MouseListener {
 	// This method checks for collisions with the sides of bricks and changes the course of the ball accordingly 
 	public static void checkSideHits(Brick tempBrick, Ball tempBall){
 		if (((tempBall.getBounds().getX()) >= (tempBrick.getBounds().getX() + tempBrick.getBounds().getWidth() - 1)) ){
-			tempBall.setXa(tempBall.getXa() * (-1));
-			tempBall.setYa(tempBall.getYa() * (-1));
+			tempBall.setXa((int) (tempBall.getXa() * (-1)));
+			tempBall.setYa((int) (tempBall.getYa() * (-1)));
 		}else if (((tempBall.getBounds().getX() + tempBall.DIAMETER) <= (tempBrick.getBounds().getX() + 1))){
-			tempBall.setXa(tempBall.getXa() * (-1));
-			tempBall.setYa(tempBall.getYa() * (-1));
+			tempBall.setXa((int) (tempBall.getXa() * (-1)));
+			tempBall.setYa((int) (tempBall.getYa() * (-1)));
 		}
 	}
 	
@@ -362,7 +362,7 @@ public class Game extends JPanel implements MouseListener {
 	
 	public Powerup generatePowerup(Brick currentBrick){
 		int tempRandNum2 = randInt(1,12); 
-		//int tempRandNum2 = 9; // Set this to a specific number to test one powerup
+		//int tempRandNum2 = 5; // Set this to a specific number to test one powerup
 		switch(tempRandNum2){
 			case 12:
 				System.out.println("Powerup Gained: " + "Extra Life");
@@ -423,9 +423,9 @@ public class Game extends JPanel implements MouseListener {
 		newbrick.getBounds().setBounds(-10, -10, 0, 0);
 		newbrick.setAlive(false);
 		newbrick = null;
-		int saveXa = saveBall.getXa();
+		int saveXa = (int) saveBall.getXa();
 		//saveBall.setXa(saveXa * (-1));
-		int saveYa = saveBall.getYa();
+		int saveYa = (int) saveBall.getYa();
 		saveBall.setYa(saveYa * (-1));
 	}
